Moved event and listschedule to model/event
Updated commands on listing events with date as well.
