+++ src/main/java/seedu/duke/controller/ControlManager.java
@@ -45,7 +45,7 @@ public class ControlManager {
             DataManager dataModel = new ModelExtractor(model, modelType).retrieveModel();
 
             if (modelType == ModelType.EVENT) {
-                new ListCommand().execute(model.getEventManager());
+                new ListCommand(userInput).execute(model.getEventManager());
             } else {
                 actionableCommand.execute(dataModel);
             }
