+++ src/main/java/seedu/duke/controller/command/ListCommand.java
@@ -4,8 +4,17 @@ import seedu.duke.model.DataManager;
 import seedu.duke.model.event.EventManager;
 
 public class ListCommand extends Command {
+    private String userInput;
+
+    public ListCommand() {
+    }
+
+    public ListCommand(String userInput) {
+        this.userInput = userInput;
+    }
+
     public void execute(EventManager eventManager) {
-        eventManager.listSchedule();
+        eventManager.listSchedule(userInput);
     }
 
     @Override
