+++ src/main/java/seedu/duke/model/event/EventManager.java
@@ -2,14 +2,12 @@ package seedu.duke.model.event;
 
 import seedu.duke.model.event.cca.EventCcaManager;
 import seedu.duke.model.event.classlesson.EventClassManager;
-import seedu.duke.ListSchedule;
 import seedu.duke.common.Messages;
 import seedu.duke.model.event.test.EventTestManager;
 import seedu.duke.model.event.tuition.EventTuitionManager;
 import seedu.duke.ui.UserInterface;
 import seedu.duke.exception.EmptyListException;
 
-import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 
@@ -18,7 +16,7 @@ public class EventManager {
     private static EventTestManager eventTestManager;
     private static EventCcaManager eventCcaManager;
     private static EventTuitionManager eventTuitionManager;
-    private UserInterface userInterface;
+    private final UserInterface userInterface;
 
     public EventManager(EventParameter eventParameter) {
         eventClassManager = new EventClassManager(eventParameter.getClasses());
@@ -44,52 +42,17 @@ public class EventManager {
         return eventTuitionManager;
     }
 
-    public void listSchedule() {
+    public void listSchedule(String userInput) {
         try {
-            ListSchedule listSchedule = new ListSchedule(
-                eventClassManager.getClasses(), eventCcaManager.getCcaList(),
-                eventTestManager.getTestList(), eventTuitionManager.getTuitions());
-            ArrayList<String> printedEvents = listSchedule.getAllEventsPrinted();
+            String dateParam = userInput.split(" ").length == 2 ? null : userInput.split(" ")[2];
+            ListSchedule listSchedule = new ListSchedule(dateParam, eventClassManager.getClasses(),
+                    eventCcaManager.getCcaList(), eventTestManager.getTestList(), eventTuitionManager.getTuitions());
+            ArrayList<String> printedEvents = listSchedule.getPrintableEvents();
             userInterface.printArray(printedEvents);
         } catch (EmptyListException e) {
             userInterface.showToUser(Messages.MESSAGE_EMPTY_SCHEDULE_LIST);
-        }
-    }
-
-    public void listDate(String[] userInput) {
-        try {
-            if (userInput[2].contains("today")) {
-                listDateInputDate(LocalDate.now().toString());
-            } else {
-                listDateInputDate(userInput[2]);
-            }
-        } catch (IndexOutOfBoundsException e) {
-            System.out.println("☹ OOPS!!! Please enter valid date and time in format yyyy-mm-dd HHMM or today!");
-        }
-    }
-
-    public void listDateInputDate(String userInput) {
-        try {
-            ListSchedule listSchedule = new ListSchedule(
-                    eventClassManager.getClasses(), eventCcaManager.getCcaList(),
-                    eventTestManager.getTestList(), eventTuitionManager.getTuitions());
-            ArrayList<String> printedEvents = listSchedule.getAllEventsPrintedDate(userInput);
-            if (printedEvents.size() > 0) {
-                printArray(printedEvents);
-            } else {
-                System.out.println("No events are due.");
-            }
-        } catch (EmptyListException e) {
-            userInterface.showToUser(Messages.MESSAGE_EMPTY_SCHEDULE_LIST);
         } catch (DateTimeParseException e) {
-            System.out.println("☹ OOPS!!! Please enter valid date and time in format yyyy-mm-dd HHMM or today!");
-        }
-    }
-
-    private static void printArray(ArrayList<String> printedEvents) {
-        assert printedEvents != null;
-        for (String line : printedEvents) {
-            System.out.println(line);
+            System.out.println("☹ OOPS!!! Please enter valid date and time in format yyyy-mm-dd or today!");
         }
     }
 }
