+++ src/main/java/seedu/duke/model/event/test/EventTestManager.java
@@ -1,6 +1,6 @@
 package seedu.duke.model.event.test;
 
-import seedu.duke.Event;
+import seedu.duke.model.event.Event;
 import seedu.duke.common.LogManager;
 import seedu.duke.common.Messages;
 import seedu.duke.exception.TestEmptyStringException;
