--- src/main/java/seedu/duke/parser/CommandType.java
@@ -1,10 +1,0 @@
-package seedu.duke.parser;
-
-/**
- * Represents all possible commands available to the user.
- */
-public enum CommandType {
-    ADD_CLASS, ADD_CCA, ADD_TEST, ADD_QUIZ, ADD_CONTACT, ADD_TUITION,
-    DELETE_CLASS, DELETE_CCA, DELETE_TEST, DELETE_QUIZ, DELETE_CONTACT,
-    DONE_TEST, DONE_CLASS, DONE_CCA, LIST_QUIZ, LIST_CONTACT, LIST_DATE, BYE, LIST, HELP
-}
