+++ src/test/java/seedu/duke/DoneTestJUnitTest.java
@@ -3,6 +3,7 @@ package seedu.duke;
 import org.junit.jupiter.api.Test;
 import seedu.duke.exception.TestEmptyStringException;
 import seedu.duke.exception.TestParamException;
+import seedu.duke.model.event.Event;
 import seedu.duke.model.event.test.EventTestManager;
 
 import java.util.ArrayList;
